requirements.txt
fastapi
uvicorn
psycopg2-binary
python-multipart
jinja2
